Default `cilium.giantswarm.io/pod-cidr` annotation on AWSCluster CR when upgrading from v17 to v18.
