+++ pkg/aws/v1alpha3/awscluster/validate_awscluster.go
@@ -3,7 +3,6 @@ package awscluster
 import (
 	"context"
 	"fmt"
-	"net"
 	"strings"
 
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v6/pkg/apis/infrastructure/v1alpha3"
@@ -12,7 +11,6 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	admissionv1 "k8s.io/api/admission/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/aws-admission-controller/v4/config"
 	aws "github.com/giantswarm/aws-admission-controller/v4/pkg/aws/v1alpha3"
@@ -22,8 +20,6 @@ import (
 type Validator struct {
 	k8sClient k8sclient.Interface
 	logger    micrologger.Logger
-
-	ipamCidrBlock string
 }
 
 func NewValidator(config config.Config) (*Validator, error) {
@@ -37,8 +33,6 @@ func NewValidator(config config.Config) (*Validator, error) {
 	v := &Validator{
 		k8sClient: config.K8sClient,
 		logger:    config.Logger,
-
-		ipamCidrBlock: config.IPAMNetworkCIDR,
 	}
 
 	return v, nil
@@ -117,11 +111,6 @@ func (v *Validator) ValidateUpdate(request *admissionv1.AdmissionRequest) (bool,
 		return false, microerror.Maskf(parsingFailedError, "unable to parse awscluster: %v", err)
 	}
 
-	err = v.Cilium(awsCluster)
-	if err != nil {
-		return false, microerror.Mask(err)
-	}
-
 	err = aws.ValidateOrganizationLabelContainsExistingOrganization(context.Background(), v.k8sClient.CtrlClient(), &awsCluster)
 	if err != nil {
 		return false, microerror.Mask(err)
@@ -156,59 +145,6 @@ func (v *Validator) ValidateUpdate(request *admissionv1.AdmissionRequest) (bool,
 	}
 
 	return true, nil
-}
-
-func (v *Validator) Cilium(awsCluster infrastructurev1alpha3.AWSCluster) error {
-	podCidr, ok := awsCluster.GetAnnotations()[annotation.CiliumPodCidr]
-	if !ok {
-		return nil
-	}
-
-	_, ciliumIPNet, err := net.ParseCIDR(podCidr)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	prefix, _ := ciliumIPNet.Mask.Size()
-	if prefix > 18 {
-		return microerror.Maskf(notAllowedError,
-			fmt.Sprintf("The CIDR from annotation `%s` is not valid, please specify a network mask which is at least `/18` or bigger, e.g. `10.0.0.0/15`", annotation.CiliumPodCidr),
-		)
-	}
-
-	_, awsPodIPNet, err := net.ParseCIDR(awsCluster.Spec.Provider.Pods.CIDRBlock)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	var ipamCidr string
-	{
-		if awsCluster.Spec.Provider.Nodes.NetworkPool != "" {
-			// Cluster is using a custom network CIDR for nodes, we need to retrieve the NetworkPool CR to know it.
-			np := infrastructurev1alpha3.NetworkPool{}
-			err = v.k8sClient.CtrlClient().Get(context.Background(), client.ObjectKey{Namespace: awsCluster.Namespace, Name: awsCluster.Spec.Provider.Nodes.NetworkPool}, &np)
-			if err != nil {
-				return microerror.Mask(err)
-			}
-
-			ipamCidr = np.Spec.CIDRBlock
-		} else {
-			ipamCidr = v.ipamCidrBlock
-		}
-	}
-
-	_, ipamIPNet, err := net.ParseCIDR(ipamCidr)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if intersect(ciliumIPNet, awsPodIPNet) || intersect(ciliumIPNet, ipamIPNet) {
-		return microerror.Maskf(notAllowedError,
-			fmt.Sprintf("The CIDR from annotation `%s` intersects with the current CIDRs `%s`, `%s`, please specify a different CIDR", annotation.CiliumPodCidr, awsCluster.Spec.Provider.Pods.CIDRBlock, ipamCidr),
-		)
-
-	}
-
-	return nil
 }
 
 func (v *Validator) AWSClusterAnnotationCNIMinimumIPTarget(awsCluster infrastructurev1alpha3.AWSCluster) error {
@@ -308,8 +244,4 @@ func (v *Validator) Log(keyVals ...interface{}) {
 
 func (v *Validator) Resource() string {
 	return "awscluster"
-}
-
-func intersect(n1, n2 *net.IPNet) bool {
-	return n2.Contains(n1.IP) || n1.Contains(n2.IP)
 }
