+++ pkg/aws/v1alpha3/awscluster/validate_awscluster_test.go
@@ -65,73 +65,3 @@ func TestAWSCNIPrefix(t *testing.T) {
 		})
 	}
 }
-
-func TestCilium(t *testing.T) {
-	testCases := []struct {
-		ctx  context.Context
-		name string
-
-		ciliumCidr    string
-		ipamCidrBlock string
-		podCidrBlock  string
-		err           error
-	}{
-		{
-			// CNI CIDR is not allowed too small.
-			name: "case 0",
-			ctx:  context.Background(),
-
-			ciliumCidr:    "10.0.0.0/25",
-			ipamCidrBlock: "10.5.0.0/16",
-			podCidrBlock:  "10.4.0.0/16",
-			err:           notAllowedError,
-		},
-		{
-			// CNI CIDR is not allowed, overlapping CIDR's.
-			name: "case 1",
-			ctx:  context.Background(),
-
-			ciliumCidr:    "10.0.0.0/8",
-			ipamCidrBlock: "10.5.0.0/16",
-			podCidrBlock:  "10.0.0.0/16",
-			err:           notAllowedError,
-		},
-		{
-			// CNI CIDR is allowed, no overlapping CIDR's.
-			name: "case 1",
-			ctx:  context.Background(),
-
-			ciliumCidr:    "10.0.0.0/16",
-			ipamCidrBlock: "10.1.0.0/16",
-			podCidrBlock:  "10.2.0.0/16",
-			err:           nil,
-		},
-	}
-	for i, tc := range testCases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			var err error
-
-			fakeK8sClient := unittest.FakeK8sClient()
-			validate := &Validator{
-				k8sClient: fakeK8sClient,
-				logger:    microloggertest.New(),
-
-				ipamCidrBlock: tc.ipamCidrBlock,
-			}
-
-			// run admission request to default AWSCluster Pod CIDR
-			awsCluster := unittest.DefaultAWSCluster()
-
-			// set CNI prefix annotation
-			awsCluster.SetAnnotations(map[string]string{
-				annotation.CiliumPodCidr: tc.ciliumCidr,
-			})
-			awsCluster.Spec.Provider.Pods.CIDRBlock = tc.podCidrBlock
-
-			err = validate.Cilium(*awsCluster)
-			if microerror.Cause(err) != tc.err {
-				t.Fatal(err)
-			}
-		})
-	}
-}
